% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_validity.R
\name{print.content_validity}
\alias{print.content_validity}
\title{Print method for content_validity objects}
\usage{
\method{print}{content_validity}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{content_validity} object returned by \code{\link[=content_validity]{content_validity()}}.}

\item{digits}{Integer. Number of digits to round numeric output to.}

\item{...}{Currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for content_validity objects
}
